Rename modules after moving them
