+++ nym-vpn-core/nym-vpnd/src/command_interface/protobuf/connection_state.rs
@@ -1,0 +1,23 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+use crate::service::VpnServiceStateChange;
+use nym_vpn_proto::{ConnectionStateChange, ConnectionStatus, Error as ProtoError};
+
+impl From<VpnServiceStateChange> for ConnectionStateChange {
+    fn from(status: VpnServiceStateChange) -> Self {
+        let mut error = None;
+        let status = match status {
+            VpnServiceStateChange::NotConnected => ConnectionStatus::NotConnected,
+            VpnServiceStateChange::Connecting => ConnectionStatus::Connecting,
+            VpnServiceStateChange::Connected => ConnectionStatus::Connected,
+            VpnServiceStateChange::Disconnecting => ConnectionStatus::Disconnecting,
+            VpnServiceStateChange::ConnectionFailed(reason) => {
+                error = Some(ProtoError::from(reason));
+                ConnectionStatus::ConnectionFailed
+            }
+        } as i32;
+
+        ConnectionStateChange { status, error }
+    }
+}
