+++ nym-vpn-core/nym-vpnd/src/command_interface/protobuf/error.rs
