+++ nym-vpn-core/nym-vpnd/src/types/mod.rs
@@ -1,4 +1,4 @@
 // Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-pub mod gateways;
+pub mod gateway;
